Move RenderNotifier from compositing to servo
